requirements.txt
ecdsa
PyNaCl
pytest
